:Fix searchparam오류 수정
